Create a list of analysis from a results table

Create a list of analysis from a results table

Usage

create_loa_from_results(results_table, analysis_key_column = "analysis_key")

Arguments

results_table

Results table with an analysis key

analysis_key_column

character vector with the name of the analysis key column. Default is "analysis key"

Value

a list of analysis

Examples

create_loa_from_results(analysistools_MSNA_template_with_ratio_results_table$results_table)
#> # A tibble: 15 × 8
#>    analysis_type        analysis_var      group_var level analysis_var_numerator
#>    <chr>                <chr>             <chr>     <dbl> <chr>                 
#>  1 prop_select_one      admin1            NA         0.95 NA                    
#>  2 mean                 income_v1_salari… NA         0.95 NA                    
#>  3 median               income_v1_salari… NA         0.95 NA                    
#>  4 mean                 expenditure_debt  NA         0.95 NA                    
#>  5 median               expenditure_debt  NA         0.95 NA                    
#>  6 ratio                NA                NA         0.95 income_v1_salaried_wo…
#>  7 prop_select_one      wash_drinkingwat… NA         0.95 NA                    
#>  8 prop_select_multiple edu_learning_con… NA         0.95 NA                    
#>  9 mean                 income_v1_salari… admin1     0.95 NA                    
#> 10 median               income_v1_salari… admin1     0.95 NA                    
#> 11 mean                 expenditure_debt  admin1     0.95 NA                    
#> 12 median               expenditure_debt  admin1     0.95 NA                    
#> 13 ratio                NA                admin1     0.95 income_v1_salaried_wo…
#> 14 prop_select_one      wash_drinkingwat… admin1     0.95 NA                    
#> 15 prop_select_multiple edu_learning_con… admin1     0.95 NA                    
#> # ℹ 3 more variables: analysis_var_denominator <chr>, numerator_NA_to_0 <lgl>,
#> #   filter_denominator_0 <lgl>